Mypy works, but tests not
