+++ checkov/common/runners/runner_registry.py
@@ -43,6 +43,7 @@ from checkov.common.util.secrets_omitter import SecretsOmitter
 from checkov.common.util.type_forcers import convert_csv_string_arg_to_list, force_list
 from checkov.sca_image.runner import Runner as image_runner
 from checkov.common.secrets.consts import SECRET_VALIDATION_STATUSES
+from checkov.terraform import TFDefinitionKey
 from checkov.terraform.context_parsers.registry import parser_registry
 from checkov.terraform.tf_parser import TFParser
 
@@ -658,7 +659,7 @@ class RunnerRegistry:
         for repo_root in repo_roots:
             parsing_errors: dict[str, Exception] = {}
             repo_root = os.path.abspath(repo_root)
-            tf_definitions = TFParser().parse_directory(
+            tf_definitions: dict[TFDefinitionKey, dict[str, list[dict[str, Any]]]] = TFParser().parse_directory(
                 directory=repo_root,  # assume plan file is in the repo-root
                 out_parsing_errors=parsing_errors,
                 download_external_modules=download_external_modules,
@@ -667,7 +668,7 @@ class RunnerRegistry:
 
         enriched_resources = {}
         for repo_root, parse_results in repo_definitions.items():
-            definitions = parse_results['tf_definitions']
+            definitions = cast(dict[TFDefinitionKey, dict[str, list[dict[str, Any]]]], parse_results['tf_definitions'])
             for full_file_path, definition in definitions.items():
                 definitions_context = parser_registry.enrich_definitions_context((full_file_path, definition))
                 abs_scanned_file = full_file_path.file_path
@@ -680,7 +681,7 @@ class RunnerRegistry:
                             entity_id = ".".join(definition_path)
                             entity_context_path = [block_type] + definition_path
                             entity_context = data_structures_utils.get_inner_dict(
-                                definitions_context[full_file_path.file_path], entity_context_path
+                                definitions_context[full_file_path], entity_context_path
                             )
                             entity_lines_range = [
                                 entity_context.get("start_line"),
