+++ checkov/terraform/context_parsers/base_parser.py
@@ -16,6 +16,7 @@ from checkov.common.models.enums import ContextCategories
 from checkov.common.resource_code_logger_filter import add_resource_code_filter_to_logger
 from checkov.common.typing import TFDefinitionKeyType
 from checkov.common.util.parser_utils import get_abs_path
+from checkov.terraform import TFDefinitionKey
 from checkov.terraform.context_parsers.registry import parser_registry
 
 OPEN_CURLY = "{"
@@ -155,7 +156,7 @@ class BaseContextParser(ABC):
         return end_line_num
 
     def run(
-            self, tf_file: TFDefinitionKeyType, definition_blocks: List[Dict[str, Any]], collect_skip_comments: bool = True
+            self, tf_file: TFDefinitionKey, definition_blocks: List[Dict[str, Any]], collect_skip_comments: bool = True
     ) -> Dict[str, Any]:
         # TF files for loaded modules have this formation:  <file>[<referrer>#<index>]
         # Chop off everything after the file name for our purposes here
