+++ checkov/terraform/context_parsers/registry.py
@@ -7,8 +7,6 @@ from typing import Dict, TYPE_CHECKING, Tuple, List, Any
 import dpath
 
 from checkov.common.resource_code_logger_filter import add_resource_code_filter_to_logger
-from checkov.common.runners.base_runner import strtobool
-from checkov.common.typing import TFDefinitionKeyType
 from checkov.terraform.modules.module_objects import TFDefinitionKey
 
 if TYPE_CHECKING:
@@ -17,7 +15,7 @@ if TYPE_CHECKING:
 
 class ParserRegistry:
     context_parsers: Dict[str, "BaseContextParser"] = {}  # noqa: CCE003
-    definitions_context: Dict[TFDefinitionKeyType, Dict[str, Dict[str, Any]]] = {}  # noqa: CCE003
+    definitions_context: Dict[TFDefinitionKey, Dict[str, Dict[str, Any]]] = {}  # noqa: CCE003
 
     def __init__(self) -> None:
         self.logger = logging.getLogger(__name__)
@@ -30,25 +28,21 @@ class ParserRegistry:
         self.definitions_context = {}
 
     def enrich_definitions_context(
-        self, definitions: Tuple[str, Dict[str, List[Dict[str, Any]]]], collect_skip_comments: bool = True
-    ) -> Dict[TFDefinitionKeyType, Dict[str, Dict[str, Any]]]:
+        self, definitions: Tuple[TFDefinitionKey, Dict[str, List[Dict[str, Any]]]], collect_skip_comments: bool = True
+    ) -> Dict[TFDefinitionKey, Dict[str, Dict[str, Any]]]:
         supported_definitions = [parser_type for parser_type in self.context_parsers.keys()]
         (tf_definition_key, definition_blocks_types) = definitions
-        enable_definition_key = strtobool(os.getenv('ENABLE_DEFINITION_KEY', 'False'))
-        if isinstance(tf_definition_key, TFDefinitionKey):
-            tf_file: TFDefinitionKeyType = tf_definition_key.file_path if not enable_definition_key else tf_definition_key
-        else:
-            tf_file = tf_definition_key if not enable_definition_key \
-                else TFDefinitionKey(file_path=tf_definition_key, tf_source_modules=None)
+        tf_file = tf_definition_key.file_path
 
         if definition_blocks_types:
             definition_blocks_types = {x: definition_blocks_types[x] for x in definition_blocks_types.keys()}
             for definition_type in definition_blocks_types.keys():
                 if definition_type in supported_definitions:
-                    dpath.new(self.definitions_context, [tf_file, definition_type], {})
+                    dpath.new(self.definitions_context, [tf_definition_key, definition_type], {})
                     context_parser = self.context_parsers[definition_type]
                     definition_blocks = definition_blocks_types[definition_type]
-                    self.definitions_context[tf_file][definition_type] = context_parser.run(tf_file, definition_blocks, collect_skip_comments)
+                    self.definitions_context[tf_definition_key][definition_type] = \
+                        context_parser.run(tf_definition_key, definition_blocks, collect_skip_comments)
         return self.definitions_context
