+++ checkov/terraform/graph_builder/graph_components/module.py
@@ -6,7 +6,8 @@ from typing import List, Dict, Any, Set, Callable, Tuple, TYPE_CHECKING, Optiona
 
 from checkov.common.runners.base_runner import strtobool
 from checkov.common.util.data_structures_utils import pickle_deepcopy
-from checkov.common.util.parser_utils import get_abs_path, get_module_from_full_path
+from checkov.common.util.parser_utils import get_abs_path
+from checkov.terraform import get_module_from_full_path
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
 from checkov.terraform.graph_builder.graph_components.blocks import TerraformBlock
 from checkov.terraform.parser_functions import handle_dynamic_values
