+++ checkov/terraform/runner.py
@@ -24,8 +24,8 @@ from checkov.common.runners.base_runner import BaseRunner, CHECKOV_CREATE_GRAPH
 from checkov.common.util import data_structures_utils
 from checkov.common.util.consts import RESOLVED_MODULE_ENTRY_NAME
 from checkov.common.util.data_structures_utils import pickle_deepcopy
-from checkov.common.util.parser_utils import get_module_from_full_path, get_abs_path, get_module_name, \
-    strip_terraform_module_referrer
+from checkov.common.util.parser_utils import get_abs_path, strip_terraform_module_referrer
+from checkov.terraform import get_module_from_full_path, get_module_name
 from checkov.common.util.secrets import omit_secret_value_from_checks, omit_secret_value_from_graph_checks
 from checkov.common.variables.context import EvaluationContext
 from checkov.runner_filter import RunnerFilter
@@ -304,8 +304,7 @@ class Runner(ImageReferencerMixin[None], BaseRunner[TerraformGraphManager]):
         block_type = entity[CustomAttributes.BLOCK_TYPE]
         full_file_path = entity[CustomAttributes.FILE_PATH]
 
-        if strtobool(os.getenv('ENABLE_DEFINITION_KEY', 'False')):
-            full_file_path = TFDefinitionKey(file_path=entity.get(CustomAttributes.FILE_PATH), tf_source_modules=entity.get(CustomAttributes.SOURCE_MODULE_OBJECT))
+        full_file_path = TFDefinitionKey(file_path=entity.get(CustomAttributes.FILE_PATH), tf_source_modules=entity.get(CustomAttributes.SOURCE_MODULE_OBJECT))
 
         definition_path = entity[CustomAttributes.BLOCK_NAME].split('.')
         entity_context_path = [block_type] + definition_path
